requirements.txt
neo4j==5.28.1
dotenv==0.9.9
gradio==5.34.2
torch==2.7.1
langchain==0.3.25
langchain-community==0.3.25
soundfile==0.13.1
git+https://github.com/openai/whisper.git
gTTS==2.5.4
pygame==2.6.1
beautifulsoup4==4.13.4
spacy==3.8.7
sentence-transformers==4.1.0
langchain-google-genai==2.1.5
faiss-cpu==1.11.0